+++ src/Compilers/CSharp/Test/EndToEnd/EndToEndTests.cs
@@ -17,6 +17,7 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Threading.Tasks;
 using System.Threading;
 using System.Diagnostics;
+using Roslyn.Test.Utilities.TestGenerators;
 
 namespace Microsoft.CodeAnalysis.CSharp.UnitTests.EndToEnd
 {
@@ -669,6 +670,56 @@ $@"        if (F({i}))
                 comp.VerifyDiagnostics();
                 Assert.Equal(localFunctions ? 20 : 40, data.LambdaBindingCount);
             }, timeout: TimeSpan.FromSeconds(5));
+        }
+
+        [Fact, WorkItem("https://github.com/dotnet/roslyn/pull/70791")]
+        public void ForAttributeWithMetadataName_DeepRecursion()
+        {
+            var deeplyRecursive = string.Join("+", Enumerable.Repeat(""" "a" """, 20_000));
+            var source = $$"""
+                class Ex
+                {
+                    void M()
+                    {
+                        var v ={{deeplyRecursive}};
+                    }
+                }
+
+                [N1.X]
+                class C1 { }
+                [N2.X]
+                class C2 { }
+
+                namespace N1
+                {
+                    class XAttribute : System.Attribute { }
+                }
+
+                namespace N2
+                {
+                    class XAttribute : System.Attribute { }
+                }
+                """;
+            var parseOptions = TestOptions.RegularPreview;
+            Compilation compilation = CreateCompilation(source, options: TestOptions.DebugDllThrowing, parseOptions: parseOptions);
+
+            Assert.Single(compilation.SyntaxTrees);
+
+            var generator = new IncrementalGeneratorWrapper(new PipelineCallbackGenerator(ctx =>
+            {
+                var input = ctx.SyntaxProvider.ForAttributeWithMetadataName(
+                    "N1.XAttribute",
+                    (node, _) => node is ClassDeclarationSyntax,
+                    (context, _) => (ClassDeclarationSyntax)context.TargetNode);
+                ctx.RegisterSourceOutput(input, (spc, node) => { });
+            }));
+
+            GeneratorDriver driver = CSharpGeneratorDriver.Create(new ISourceGenerator[] { generator }, parseOptions: parseOptions, driverOptions: new GeneratorDriverOptions(IncrementalGeneratorOutputKind.None, trackIncrementalGeneratorSteps: true));
+            driver = driver.RunGenerators(compilation);
+            var runResult = driver.GetRunResult().Results[0];
+
+            Assert.Collection(runResult.TrackedSteps["result_ForAttributeWithMetadataName"],
+                step => Assert.True(step.Outputs.Single().Value is ClassDeclarationSyntax { Identifier.ValueText: "C1" }));
         }
     }
 }
